Passwd: support nis [+-]username syntax
The /etc/passwd file can contain nis users with the syntax [+-]username for
the pruprose of including or excluding specific accounts.

Importing of particular users' accounts can be restricted by using the
prefix "+" with a username.

    root:x:0:0:root user:/home/root:/bin/bash
    +bob::::::
    +alice::::Alice:/home/alice:/bin/bash

Certain users' accounts can be excluded from being imported by using the
"-" prefix with a username.

    root:x:0:0:root user:/home/root:/bin/bash
    -bob::::::
    +

The string that can comprise a valid username has been updated to cover a
range slightly larger that the definitions in POSIX
(http://pubs.opengroup.org/onlinepubs/9699919799/ , section 3.431),
is_valid_name() in chkname.c in shadow-utils and the NAME_REGEX definition
in distributions where defined. Rx.word was too broad and did not allow for
the -username syntax.