Grub: support a password stanza inside a boot/title section
The biggest change in this commit is changing the structure of lns: it now
says that boot sections have to come after the general menu_setting and
debian sections. Since 'password' is legal in both a top-level menu_setting
and in a boot section, that is necessary to keep the lens unambiguous.

Fixes https://github.com/hercules-team/augeas/issues/229