Create change
GitHub Pull Request: https://api.github.com/repos/hercules-team/augeas/pulls/244

RFC: Add a 'region' lens to make controlling spacing easier

This is not meant to be committed yet, but I would really appreciate feedback based on people's experimenting with these patches.



So far, how deleted text is handled when entries are added to or removed

from the tree is tightly bound up with how the tree was structured,

because the subtree combinator [ .. ] controlled both: construction of new

tree nodes and glomming deleted stuff together.



In addition, the [ .. ] does the glomming based on the key of the tree node

it constructs, so that deleting a node from hte middle of a number of nodes

with the same key had the weird effect that entries tended to 'shift up'.



The region combinator tries to address this, in that all it does is

demarcate how deleted stuff should be glommed, and it does that based on

the position in the input file, not on the value of a key.



Syntactically, this is written as < l >, though other syntax, like 'region

l', or '<< l >>' might be better.



It would be great if this could also be used to indent new things the same

as surrounding stuff, i.e. if we have a lens < l >* and we add en entry in

the middle of it, if that would automatically mean that the new entry gets

the indentation of one of the surrounding lines - that's not implemented

yet.



I would really like to see and hear how this does or does not address some

of the spacing/indentation problems that people have been having, ideally

demonstrated by a simple test.

Patch-set: 1
Change-id: I70100c6c4522212a06306db0cd72113783a70fbd
Subject: RFC: Add a 'region' lens to make controlling spacing easier
Branch: refs/heads/master
Commit: d38e7116c42c38365fb1f18bca81a7f93fb2f44d
Groups: d38e7116c42c38365fb1f18bca81a7f93fb2f44d