* test-perf.c (testPerfPredicate): add test showing slow predicates
When parsing a file with NagiosObjects and many objects inside, a large
tree forms like this:

    /files/etc/nagios/objects/services.cfg/service[1]
    /files/etc/nagios/objects/services.cfg/service[2]
    ...
    /files/etc/nagios/objects/services.cfg/service[5000]

aug_get performance is pretty terrible at this size when using a path
above, due to the evaluation of the [5000] predicate.  Paths without a
predicate (e.g. using seq or uniquely labelled nodes) are retrieved
quickly.

The time taken to set and get 5,000 such nodes is recorded in
tests/test-perf.log and is currently about:

    testPerfPredicate = 66636ms

A helper to run Valgrind's callgrind tool is added to src/try.