Speed up the duplicate check when constructing nodesets
When we evaluate path expressions, we need to construct nodesets, and they
truly need to be sets, i.e. contain each node at most once. This leads to
problems when a node has lots of children with the same name. The duplicate
check in ns_add lead to O(n^2) behavior for nodes with n children.

We simplify this by adding a flag to each tree node that we use only around
repeated calls of ns_add when we construct a node set. This is possible
since the construction of nodesets is very local, and it is therefore easy
to determine where we need to put the call to ns_clear_added when we are
done building the nodeset.