* src/pathx.c (ns_from_locpath): special-case 'name[42]' path expressions
We know that in a path expression like 'name[42]', only one node can
possibly match, and that we simply need to step through all the nodes and
count until we've reached the 42nd matching node. This greatly simplifies
how we construct the resulting nodeset, and is done in the new function
position_filter.