Skip hidden tree nodes in path expressions
When we were evaluating path expressions, we were not skipping hidden tree
nodes, i.e. tree nodes whose label is NULL. In most cases, that's not
visible to users, but when we match something like /foo/*[2] where /foo has
a bunch of hidden children, the result was not what the user might expect,
as we might actually return a hidden node to aug_match, which would
suppress that node and make it seem to the user as if nothing had matched.

With this patch, we try to skip over hidden nodes while evaluating the path
expression to avoid this issue.

Thanks to Xavier Mol for reporting this