Fix buffer overflow in format_union_atype
The change in commit 234de0e did not correctly calculate the length of the
output buffer, which would lead to writing beyond the end of that buffer,
and ultimately to memory corruption.

The format_atype (and ultimately, format_union_atype) functions are not
just used for debugging printing purposes, but also for error reporting
when users create invalid subtrees and try to save them. In that case,
merely creating an invalid tree can lead to a segfault.

This change corrects the length calculation and adds a test that fails with
the old length calculation, provided the test is run with MALLOC_CHECK_=3
in the environment.

Fixes issue #349