* src/put.c (skel_instance_of): fix illegal memory access
We used to always access skel->skels; but that is only valid if the tag for
skel->skels is L_CONCAT or L_STAR. Because of this, we could, for example
access skel->skels for a skel tagged as L_DEL, which is illegal.

This patch adds a fix to the L_CONCAT case so that we only look at
skel->skels if the skel is tagged that way.

Fixes https://github.com/hercules-team/augeas/issues/398