* src/get.c (visit_enter, visit_exit): fix use-after-free when spans are used
When spans are enabled, we used to store a a pointer to the same span both
in a frame and in the current state for L_SUBTREE lenses, which meant that
two different tree nodes might have pointers to the same span. That is
counter to the assumptions made about tree nodes, in that they need to own
the spans attached to them, resulting in a use-after-free. It would also
record incorrect span information in one of the tree nodes.  We now
allocate a new span when we store one in the frame - that way we are
guaranteed that there is no aliasing.

With L_MAYBE lenses, visit_enter allocated a new span in state, thus
possibly leaking the one that might be there already. We now store that
span in the frame that we push for L_MAYBE, avoiding that leak.

Finally, we popped frames in visit_exit without doing anything with
them. We now also free any associated span to make sure we do not leak that
information.

Fixes https://github.com/hercules-team/augeas/issues/397