* src/get.c: simplify recursive lens handling by fixing frame stack
The stack for frames on which we remember results from traversing the parse
tree with visit_enter/visit_exit was implemented very oddly: pop_frame
removed a frame and returned the new top. It now returns the top of the
stack that we just popped, i.e., the old top, which makes it possible to
get rid of some odd ways in which we got intermediate results off the
stack.