Cron_User: new lens for user crontab files
The user crontab files in /var/spool/cron have a slightly different format
from /etc/crontab in that they do not have a 'user' field as that's implied
by the filename. Unfortunately, it is not possible to use Cron.lns, since
defining Cron.entry as

  let entry       = [ label "entry" . indent
                    . prefix?
                    . ( time | schedule )
                    . (sep_spc . user)?
                    . sep_spc . store Rx.space_in . eol ]

(i.e., with a ( .. )? around the user name) leads to typecheck errors.

This lens is a minor variation on Cron.lns and just removes the bits about
the user field.