Properly check composite regular expressions for syntax errors
We used to not properly check the syntax of regular expressions that are
constructed. For example, in the construct

    let rx = /a/ | /b)/

we would not check whether /b)/ is syntactically correct, which would later
lead to Augeas crashing. The reason for this was that literals had types
assigned to them before we actually checked them, i.e., the assumption was
that values didn't need any checking beyond setting their type, which is
not true for regular expressions.