put.c: track the fully qualified path of our position in the tree
Previously, we only tracked a path like /a/b/c and printed that in errors
to indicate where in the tree the error occurred. That was ambiguous when
there were multiple /a/b's for example. We now include the more precise
path /a[2]/b[3]/c in error messages to clarify that.