Remove flags field from struct info
The flags field in struct info was only used to pass whether spans are
enabled to lns_get. Instead of having this field in every struct info, we
just pass it into lns_get explicitly shich puts it into its struct
state. That saves a small amount of memory as it makes struct info, of
which we have a lot, 4 bytes smaller.