* src/fa.c (fa_make_empty): do not mark fa as minimal or deterministic
Even though the empty fa is both minimal and deterministic, this function
is also used to allocate new fa's which may be modified further,
invalidating their being marked as minimal or deterministic. For example,
the fa constructed by fa_intersect was erroneously marked as minimal
because of this, making it impossible to actually minimize that fa.

Not marking the empty fa as minimal and deterministic will not cause any
problems, and may only cause us to unnecessarily try to determinize or
minimize the fa, which, given the size of the empty fa will be very fast
operations.