Make sure we don't load a module but then later can't find it
If a module name contains a '/', it can happen that we find the
module (like 'dist/Fstab.lns'), load it and then later can't find it under
the invalid module name containing the slash, leading to an endless loop of
loding the same module over and over again.

We avoid this by making sure there is no '/' in the module name before
looking for the module file. Other forms of invalid module names should not
cause similar problems as we will simply fail to find the corresponding
file.

Fixes https://github.com/hercules-team/augeas/issues/522