Systemd: explicitly exclude *.d and *.wants directories
We used to get confused when writing back systemd changes and tried to
write a whole *.d or *.wants directory as one file, which would fail. By
explicitly not matching them, we make sure that we write to the files in
those directories, not the directories themselves.

Fixes https://github.com/hercules-team/augeas/issues/548