Httpd: accept whitespace comment after tag opening a section
We used to only deal with completely empty comments after a tag opening a
section, for example '<If cond>#\n', but not with a comment that consisted
only of whitespace like '<If cond># \n'. We now accept both cases,
resulting in the same tree.

Fixes https://github.com/hercules-team/augeas/issues/577