examples: turn them as standalone
Examples are supposed to be small test applications (examples, indeed)
that help users to know how the library works.  Hence, it is a good
thing that users can pick them out of the sources, and build them
standalone on their own, using the system augeas library.

Make this situation possible:
- do not rely on gnulib: they are test applications of augeas, not
  supposed to be the most perfect C examples; besides, gnulib is not
  available out of the augeas build tree
- stop including config.h: the content of the examples is not supposed
  to change depending on the augeas build configuration
- include the augeas headers as system headers, not as local ones; the
  build system takes care that the local augeas headers are used when
  building the examples as part of the sources