Fixes missing recognition of double quoted filenames. (#611)
Fixes missing recognition of double quoted filenames.

Logrotate is able use quote filenames. This by itself is fixed there. Some
distributions files are using that. Logrotate would also be able to use files
with spaces (when quoted), fixing is too complicated for a quick fix and this
should be also uncommon in real life.