Mke2fs: refactor fs_entry options
According to the mke2fs.conf(5) documentation, all the entries that
appear in the various tags in the [fs_types] stanza may be used also in
the [default] section; hence:
- fold fs_types_entry into common_entry
- move almost all the options (except fs_type, and undo_dir) from
  defaults_entry to common_entry
so the options apply to both cases.

Also, in an attempt to make the list of options slightly maintainable,
add few lists with names of options with common types (list, int, bool).