Few improvements to the Mke2fs lens (#642)
* Mke2fs: fix handling of booleans

Add a proper regexp to represent them, according to the mke2fs.conf(5)
documentation.  Fix the lazy_itable_init, and auto_64-bit_support keys
of fs_types to use this regexp, as they are parsed as boolean.

* Mke2fs: refactor fs_entry options

According to the mke2fs.conf(5) documentation, all the entries that
appear in the various tags in the [fs_types] stanza may be used also in
the [default] section; hence:
- fold fs_types_entry into common_entry
- move almost all the options (except fs_type, and undo_dir) from
  defaults_entry to common_entry
so the options apply to both cases.

Also, in an attempt to make the list of options slightly maintainable,
add few lists with names of options with common types (list, int, bool).

* Mke2fs: add more defaults/fs_types entries

Add more common entries for the [defaults] stanza, or inside a
[fs_types] tag, fixing the types to match documentation/code.

This does not cover all the possible entries available though.

* Mke2fs: handle the [options] stanza

Handle the two documented tags of it.