postgresql.aug: Allow hyphen '-' in values that don't require quotes (#700)
Since PostgreSQL 10 password_encryption is enum.
If the value is set to scram-sha-256 (which contains hyphens),
saving postgresql.conf with augeas fails with:

saving failed (run 'errors' for details)
Error in /var/lib/pgsql/13/data/postgresql.conf:782.0 (parse_skel_failed)
  Iterated lens matched less than it should
  Lens: /usr/share/augeas/lenses/dist/postgresql.aug:69.10-.46:
    Last matched: /usr/share/augeas/lenses/dist/postgresql.aug:30.10-.46:
    Next (no match): /usr/share/augeas/lenses/dist/quote.aug:117.2-.35:

This is because hyphen wasn't included in the definition of string
that doesn't require quoting.